pothole-ai-system/hardware/uno_q/python/requirements.txt
python-dotenv
moorcheh-sdk
requests